requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.26.0
cohere==4.47
pydantic==2.6.1
python-multipart==0.0.9 